FIPS compliant algorithms fix
On Windows XP and later, When the security setting "System cryptography: Use FIPS compliant algorithms for encryption, hashing, and signing" is turned on, instantiating a SHA1Managed will result in a runtime error.  

Using SHA1.Create() allows the runtime to select the appropriate algorithm.  If the setting is turned on, the FIPS-compliant native algorithm will be used, otherwise the SHA1Managed will be used.

More info: https://support.microsoft.com/en-us/kb/811833