Optimizations for test case serialization.
Introduced two flags to control SourceInformation and Serialization of xunit test case
as a TestProperty in VSTestCase object. Serialization is enabled for Discovery always and
enabled for Run All or Run Specific test operation only if the test run is triggered from
an IDE.

Moved Discovery (in case of Run All flow) to use the same XunitFrontController as execution
of an assembly. This removed the need for Serialize of XunitTestCase in Discovery[execution]
followed by a Deserialization in actual Execution.

Since the ITestCase objects are MarshalByRef, optimized the code in VsExecutionSink to lookup
a VSTestCase based on ITestCase.UniqueId instead of entire ITestCase objects.