Target frameworks not matching
Nuget.Frameworks started returning DotNetFrameworkName as "net5.0"
which is what vstest will pass down as the target framework name.

There is some guidance how to do that better, but this is a quick fix
to be able to run xUnit tests on net5.0-preview8 and newer, before we
figure out the best way to do this.